WIP -- clean up build config
Document and move around a few of the items. Also remove stuff
that was erroneously copied to the target root dir. The softwareupdater
URL, nodemanager version etc. should be checked using the files in
the actual `seattle_repy` dir that will be used for building, and not
in copies of these files that go unused otherwise.

*Note: These checks are not yet fixed in `rebuild_base_installers`!*